ToonOutlinePassNode: Refactor code.
